pyth: Print error and return 1 on mapping key failure.
Previously, pyth would return a 0 exit code but not return any data.
This could happen if the mapping key is missing or invalid.

Also, only try updating the price::pub_idx_ if the manager actually
has a publish public key. This was causing a crash when running the
tools without a publish keypair. This should be valid for running
in monitoring / consuming modes.