Update blockdaemon nginx config
When starting nginx with the standalone-dev-nginx.conf, with the
previous values in the blockdaemon config, nginx errors out with an
error such as:

    2021/09/16 11:52:29 [emerg] 126854#0: host not found in upstream "YOUR_SERVER-solana-01.bdnodes.net" in blockdaemon-virt-server-config.conf:64

Setting it to a variable delays the name resolution until a request
is made. This makes it much easier to get a test nginx proxy up
and running.