Remove redundant usage of drv1_ field.
This is always 1 for all currently deployed price accounts which have had upd_ema
ran once, which covers all accounts. Therefore the condition inside upd_ema will always
be true. We can therefore remove the usage of drv1_, freeing it up to be used for
other 64-bit data in the future.